Sync person by person
This is a major change. Instead of fetching all of the events, grouping
them by person, then doing all of the syncing, we flip the logic to
fetch all of the people, then for each one, find their events and sync.

Doing it this way makes the code easier to follow. A person owns their
syncing operation instead of needing an orchestrator process.

The downside of this approach is that we make a lot more API calls. We
don't need this to be very performant, as it runs once per day, but we
will need to handle rate limiting properly.