Handle rate limiting during pagination
Due to the way automatic pagination works, if we are rate limited
between page requests, the Breathe client ends up with a nil response
body which it attempts to concatenate with the previous pages. We catch
that error, but it would be better to make the Breathe client understand
rate limits itself instead.