Match emails between systems using aliases
Some people have different email addresses associated with different
systems. This means they can't be matched, and their events can't be
synced.

This change introduces the concept of email aliases. By providing a list
of email addresses that represent the same person, we can link accounts
across systems.

We currently assume that people only have a single account on a given
system, so we don't attempt to work out _which_ account we want if
multiple match.