Output one file per person
Output one file per person, as soon as we fetched the events for that person.

Write data dumps to a gitignored folder.

In order for these data files to be more easily usable for debugging, take advantage
of the decision to output a file per person and add more data about that person, so
we can have a reproducible attempt to import this data into Productive.

Co-authored-by: Lorna <lorna@dxw.com>