Partial fix for trainings not being imported
We’ve added some debug logging for skipped trainings, and discovered
that the reason they were skipped was that start and end date attributes
are different for trainings than what we were assuming from the
structure of leave and sickness events.

The half day logic is also different, to be fixed in a future commit.