Rake task to do a dry run syncing data from files
In order to investigate issues with synchronising data from BreatheHR
into Productive, we’re adding a task to simulate a dry run using data
obtained from a data dump. This is so that we don’t need API access to
BreatheHR, as long as we have the data from someone with access.

Productive API access is less sensitive, and can be given to more team
members.