Fix rake task instructions
There was a misunderstanding about sending arguments to a rake task. It
turns out that they don’t support keyword arguments after all, so the
task was incorrectly using `emails:` as part of the first email.

Since the arguments are positional, we can’t make both optional and have
them sent in whichever combination, so remove the “optional” from the
instructions.

The next fix will be to enforce the presence of emails.