Synchronise specific records
By default, the sync task processes all the people records in BreatheHR.
This takes a long time, and it is impractical to run when there are only
a small number of accounts affected by syncronisation issues.

The sync task now looks for an environment variable, `EMAILS`, expected
to contain the comma-separated emails of the records to be processed.