Update Productive holiday event ID
Productive recently made changes to its features related to time off.

We have 5 categories of time off: Holiday, Manual override, Other leave, Regular non-working time, and Sickness.

A time off category can now be paid/unpaid and limited/unlimited. Productive have set Holiday to be limited, and all the others to be unlimited. This is not something we can edit, and there’s no explanation for why not.

If a category is unlimited, everyone has an ‘entitlement’ to it. If it’s limited, we have to add their entitlement on a per person basis.
Since Holiday is limited and can’t be made unlimited, we would have to add it to everyone’s account. This is massive pain.

My solution is to create a new time off category called `2 Holiday 2 Relaxing` with ID `81131` and have the this sync use that instead. This PR just updates the example env with the new ID, the 'real' change is to the config on Heroku.