WIP Fix after/start_date param
The redacted API's `list` method takes the param `after` and translates
it to `start_date`. We were still passing it to the redacted API as
`start_date`

ISSUE: we still get training before the start date, for some reason,
but absences and sicknesses are fine

QUESTION: do we need to support the original API, in which case do we
need to make the key conditional on whether that environment variable
is set?